+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased]
 
+## [1.4.0] - 2026-02-14
+
 ### Changed
 
 - Update default nightly toolchain to `nightly-2026-02-12`
@@ -303,7 +305,8 @@ _This release does not make any breaking changes. The version is bumped to `1.0.
 ## [0.1.0] - 2025-07-09
 
 <!-- next-url -->
-[Unreleased]: https://github.com/bluurryy/cargo-insert-docs/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/bluurryy/cargo-insert-docs/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/bluurryy/cargo-insert-docs/releases/tag/v1.4.0
 [1.3.0]: https://github.com/bluurryy/cargo-insert-docs/releases/tag/v1.3.0
 [1.2.0]: https://github.com/bluurryy/cargo-insert-docs/releases/tag/v1.2.0
 [1.1.0]: https://github.com/bluurryy/cargo-insert-docs/releases/tag/v1.1.0